+++ cryptol-remote-api/test-cryptol-remote-api.py
@@ -9,7 +9,7 @@ port = int(sys.argv[3])
 if connType == 'socket':
     c = cryptol.connect(argo.RemoteSocketProcess(host, port=port, ipv6=False))
 elif connType == 'http':
-    c = cryptol.CryptolConnection(argo.ServerConnection(argo.HttpProcess(url="http://%s:%d/" % (host,port))))
+    c = cryptol.connect(url="http://%s:%d/" % (host,port))
 else:
     raise Exception('specify socket or http for connection type')