+++ src/Cryptol/Backend/WordValue.hs
@@ -509,10 +509,12 @@ unwindThunks m =
 shiftWordByInteger ::
   Backend sym =>
   sym ->
-  (SWord sym -> SWord sym -> SEval sym (SWord sym)) ->
-  (Integer -> Integer -> Maybe Integer) ->
-  WordValue sym ->
-  SInteger sym ->
+  (SWord sym -> SWord sym -> SEval sym (SWord sym)) 
+    {- ^ operation on word values -} ->
+  (Integer -> Integer -> Maybe Integer)
+    {- ^ reindexing operation -} ->
+  WordValue sym {- ^ word value to shift -} ->
+  SInteger sym {- ^ shift amount, assumed to be in range [0,len] -} ->
   SEval sym (WordValue sym)
 
 shiftWordByInteger sym wop reindex x idx =
@@ -532,27 +534,18 @@ shiftWordByInteger sym wop reindex x idx =
       isReady sym packed >>= \case
         Just w -> shiftWordByInteger sym wop reindex (WordVal w) idx
         Nothing ->
-          case integerAsLit sym idx of
-            Just j -> bitmapWordVal sym n =<< shiftOp bs0 j
-            Nothing ->
-              do (numbits, idx_bits) <- enumerateIntBits' sym n idx
-                 bitmapWordVal sym n =<<
-                   barrelShifter sym (iteBit sym) shiftOp (Nat n) bs0 numbits (map BitIndexSegment idx_bits)
-
- where
-   shiftOp vs shft =
-      pure $ indexSeqMap $ \i ->
-        case reindex i shft of
-          Nothing -> pure $ bitLit sym False
-          Just i' -> lookupSeqMap vs i'
+          bitmapWordVal sym n =<<
+            shiftSeqByInteger sym (iteBit sym) reindex (pure (bitLit sym False)) (Nat n) bs0 idx
 
 
 {-# INLINE shiftWordByWord #-}
 shiftWordByWord ::
   Backend sym =>
   sym ->
-  (SWord sym -> SWord sym -> SEval sym (SWord sym)) ->
-  (Integer -> Integer -> Maybe Integer) ->
+  (SWord sym -> SWord sym -> SEval sym (SWord sym))
+    {- ^ operation on word values -} ->
+  (Integer -> Integer -> Maybe Integer)
+    {- ^ reindexing operation -} ->
   WordValue sym {- ^ value to shift -} ->
   WordValue sym {- ^ amount to shift -} ->
   SEval sym (WordValue sym)
@@ -572,19 +565,8 @@ shiftWordByWord sym wop reindex x idx =
       isReady sym packed >>= \case
         Just w -> shiftWordByWord sym wop reindex (WordVal w) idx
         Nothing ->
-          case idx of
-            WordVal (wordAsLit sym -> Just (_,j)) ->
-              bitmapWordVal sym n =<< shiftOp bs0 j
-            _ ->
-              do idx_segs <- enumerateIndexSegments sym idx
-                 bitmapWordVal sym n =<< barrelShifter sym (iteBit sym) shiftOp (Nat n) bs0 n idx_segs
-
- where
-   shiftOp vs shft =
-      pure $ indexSeqMap $ \i ->
-        case reindex i shft of
-          Nothing -> pure $ bitLit sym False
-          Just i' -> lookupSeqMap vs i'
+          bitmapWordVal sym n =<<
+            shiftSeqByWord sym (iteBit sym) reindex (pure (bitLit sym False)) (Nat n) bs0 idx
 
 
 {-# INLINE updateWordByWord #-}
@@ -651,17 +633,24 @@ updateWordByWord sym dir w0 idx bitval =
 shiftSeqByWord  ::
   Backend sym =>
   sym ->
-  (SBit sym -> a -> a -> SEval sym a) ->
-  (Integer -> Integer -> Maybe Integer) ->
-  SEval sym a ->
-  Nat' ->
-  SeqMap sym a ->
-  WordValue sym ->
+  (SBit sym -> a -> a -> SEval sym a)
+     {- ^ if/then/else operation of values -} ->
+  (Integer -> Integer -> Maybe Integer)
+     {- ^ reindexing operation -} ->
+  SEval sym a  {- ^ zero value -} ->
+  Nat' {- ^ size of the sequence -} ->
+  SeqMap sym a {- ^ sequence to shift -} ->
+  WordValue sym {- ^ shift amount -} ->
   SEval sym (SeqMap sym a)
 shiftSeqByWord sym merge reindex zro sz xs idx =
-  do idx_segs <- enumerateIndexSegments sym idx
-     barrelShifter sym merge shiftOp sz xs (wordValueSize sym idx) idx_segs
- where
+  wordValAsLit sym idx >>= \case
+    Just j -> shiftOp xs j
+    Nothing ->
+      do idx_segs <- enumerateIndexSegments sym idx
+         barrelShifter sym merge shiftOp sz xs idx_bits idx_segs
+  where
+   idx_bits = wordValueSize sym idx
+
    shiftOp vs shft =
      pure $ indexSeqMap $ \i ->
        case reindex i shft of